pyproject.toml
[tool.poetry]
name = "playground-datalake"
version = "0.1.0"
description = "playground project to implement and trial different data engineering concepts"
authors = ["Amir <amir0akhavan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
pycoingecko = "^3.1.0"
boto3 = "^1.35.54"
dbt-postgres = "1.7.13"
dbt-core = "1.7.13"
apache-airflow = "^2.10.3"
apache-airflow-providers-amazon = "^9.1.0"
pandas = "^2.2.3"
apache-airflow-providers-postgres = "^5.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
